Allow different kind of messages + split loading from instantiation
Breaking change that modifies the API the plugins must implement, the
RequestHandler is more general and now is only a `Handler`.
Loading plugins is also improved as the load function returns a
factory function used to instantiate plugins outside of the loading
phase and loaders are expected to cache any external resources so
instantiating becomes cheaper.